requirements.txt
gradio
google.generativeai
requests
python.dotenv